Move `Modifiers` to proper package
