+++ src/ru/nsu/ccfit/zuev/osu/game/GameEffect.java
@@ -7,7 +7,7 @@ import com.reco1l.andengine.Anchor;
 import com.reco1l.osu.Execution;
 import com.reco1l.andengine.sprite.AnimatedSprite;
 import com.reco1l.andengine.sprite.ExtendedSprite;
-import com.reco1l.osu.Modifiers;
+import com.reco1l.andengine.Modifiers;
 import com.reco1l.andengine.modifier.UniversalModifier;
 
 import org.anddev.andengine.entity.scene.Scene;
