+++ src/ru/nsu/ccfit/zuev/osu/game/GameplayModernSpinner.java
@@ -4,7 +4,7 @@ import android.graphics.PointF;
 
 import com.reco1l.osu.Execution;
 import com.reco1l.andengine.sprite.ExtendedSprite;
-import com.reco1l.osu.Modifiers;
+import com.reco1l.andengine.Modifiers;
 import com.reco1l.andengine.Anchor;
 import com.rian.osu.beatmap.hitobject.Spinner;
